% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{aic}
\alias{aic}
\alias{aicc}
\alias{bic}
\title{Information criteria}
\usage{
aic(k, log_likelihood)

aicc(n, k, log_likelihood)

bic(n, k, log_likelihood)
}
\arguments{
\item{k}{number of estimated parameters.}

\item{log_likelihood}{the maximized log-likelihood.}

\item{n}{sample size.}
}
\value{
a single numeric criterion value.
}
\description{
Information criteria
}
\examples{
aic(k = 2, log_likelihood = -121.01131220612)
aicc(n = 30, k = 2, log_likelihood = -121.01131220612)
bic(n = 30, k = 2, log_likelihood = -121.01131220612)
}
