% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula.R
\name{copula_inverse_cdf}
\alias{copula_inverse_cdf}
\title{Copula inverse CDF}
\usage{
copula_inverse_cdf(cop, u, v)
}
\arguments{
\item{cop}{a \code{corehydro_copula}.}

\item{u, v}{numeric scalars in \verb{(0, 1)}.}
}
\value{
a numeric vector of length two.
}
\description{
Copula inverse CDF
}
\examples{
copula_inverse_cdf(copula("Clayton", theta = 2), 0.3, 0.7)
}
