% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula.R
\name{copula_names}
\alias{copula_names}
\title{List the supported copula families}
\usage{
copula_names()
}
\value{
a character vector of the eight bivariate copula family names.
}
\description{
List the supported copula families
}
\examples{
copula_names()
}
