% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula.R
\name{copula_tail_dependence}
\alias{copula_tail_dependence}
\title{Copula tail dependence}
\usage{
copula_tail_dependence(cop)
}
\arguments{
\item{cop}{a \code{corehydro_copula}.}
}
\value{
a named numeric of length two, \code{lower} and \code{upper}.
}
\description{
The lower and upper tail dependence coefficients of a \code{\link[=copula]{copula()}}.
}
\examples{
copula_tail_dependence(copula("Clayton", theta = 2))
}
