% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{dft}
\alias{dft}
\title{Discrete Fourier transform}
\usage{
dft(x, inverse = FALSE)
}
\arguments{
\item{x}{numeric vector, complex data packed as \verb{[re0, im0, re1, im1, ...]}.}

\item{inverse}{if \code{TRUE}, computes \code{n} times the inverse transform (matching the C# doc
comment -- divide by \code{n} for the true inverse).}
}
\value{
the transformed vector, same length as \code{x}.
}
\description{
Mirrors the C# \code{Fourier.FFT}: an in-place complex FFT on data packed as
\verb{[re0, im0, re1, im1, ...]}; \code{length(x) / 2} must be a power of two.
}
\examples{
dft(c(1, 0, 2, 0, 3, 0, 4, 0))
}
