% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{link}
\alias{link}
\alias{link_inverse}
\alias{link_derivative}
\title{Evaluate a link function}
\usage{
link(l, x)

link_inverse(l, eta)

link_derivative(l, x)
}
\arguments{
\item{l}{a \code{corehydro_link} from \code{\link[=link_function]{link_function()}}.}

\item{x, eta}{numeric vectors on the data and linear-predictor scales.}
}
\value{
a numeric vector the same length as the input.
}
\description{
Evaluate a link function
}
\examples{
l <- link_function("Logit")
link(l, c(0.1, 0.5, 0.9))
}
