% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvdist.R
\name{mvdist_dirichlet}
\alias{mvdist_dirichlet}
\title{Construct a Dirichlet distribution}
\usage{
mvdist_dirichlet(alpha)
}
\arguments{
\item{alpha}{numeric vector of positive concentration parameters.}
}
\value{
a \code{corehydro_mvdist} of family \code{"Dirichlet"}.
}
\description{
Mirrors the C# \code{Dirichlet} class of the Numerics library.
}
\examples{
mvdist_pdf(mvdist_dirichlet(c(2, 3, 4)), c(0.2, 0.3, 0.5))
}
