% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvdist.R
\name{mvdist_mahalanobis}
\alias{mvdist_mahalanobis}
\title{Mahalanobis distance}
\usage{
mvdist_mahalanobis(mv, x)
}
\arguments{
\item{mv}{a \code{corehydro_mvdist} of family \code{"MultivariateNormal"} or \code{"MultivariateStudentT"}.}

\item{x}{numeric vector, the evaluation point.}
}
\value{
a single numeric.
}
\description{
Mahalanobis distance
}
\examples{
mvdist_mahalanobis(mvdist_normal(c(0, 0), diag(2)), c(1, 1))
}
