% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvdist.R
\name{mvdist_names}
\alias{mvdist_names}
\title{List the supported multivariate distribution families}
\usage{
mvdist_names()
}
\value{
a character vector of the five multivariate distribution family names.
}
\description{
List the supported multivariate distribution families
}
\examples{
mvdist_names()
}
