% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{link_names}
\alias{link_names}
\title{Available link function types}
\usage{
link_names()
}
\value{
a character vector of the twelve \code{type} names \code{\link[=link_function]{link_function()}} accepts.
}
\description{
Available link function types
}
\examples{
link_names()
}
